pyproject.toml
[project]
name = "ai-doctor-crew"
version = "0.1.0"
description = "AI Doctor Crew - Medical Document Processing and Analysis System with Milvus Vector Database"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.141.0",
    "crewai-tools>=0.51.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "pymilvus>=2.3.0",
    "numpy>=1.24.0",
    "google-generativeai>=0.3.0",
    "pymupdf>=1.26.3",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "pathlib2>=2.3.0",
    "pypdf>=5.8.0",
    "pypdf2>=3.0.1",
    "chromadb>=0.5.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crew"]

[tool.crewai]
agents_config = "src/crew/config/agents.yaml"
tasks_config = "src/crew/config/tasks.yaml"
